Add new test case file and remove tests from intergrated test file
